+++ src/pages/jihwan/Main/Comment.scss
@@ -1,0 +1,84 @@
+.feedWrap {
+  border: 1px solid #e6e6e6;
+  background-color: white;
+
+  .feedSubImg {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    .feedHeart,
+    .feedComment,
+    .feedUpload {
+      width: 30px;
+      margin-top: 5px;
+      padding-left: 10px;
+    }
+    .feedBookmark {
+      width: 25px;
+      margin-left: auto;
+      margin-right: 10px;
+      margin-top: 5px;
+    }
+  }
+  .contents {
+    flex-direction: column;
+    margin-left: 10px;
+    margin-right: 10px;
+    margin-top: 12px;
+
+    .contentImg {
+      width: 18px;
+      height: 18px;
+      border-radius: 100%;
+    }
+
+    .spanBold {
+      width: 100%;
+      .spBold1 {
+        font-weight: bold;
+      }
+      .spBold2 {
+        font-weight: bold;
+      }
+    }
+
+    .contentSpanWrap {
+      margin-top: 5px;
+
+      .contentSpanNBold {
+        .contentBold {
+          margin-right: 3px;
+          font-weight: bold;
+        }
+      }
+    }
+
+    .comment {
+      display: flex;
+      justify-content: space-between;
+      border-top: 1px solid #e6e6e6;
+      margin-top: 10px;
+
+      .commentInput {
+        width: 90%;
+        margin-top: 7px;
+        margin-bottom: 7px;
+        border: none;
+      }
+
+      .registerBtn {
+        background-color: white;
+        color: #c4e1fb;
+        border: none;
+      }
+      .registerBtnActive {
+        background-color: skyblue;
+        color: white;
+        font-weight: bold;
+        border: none;
+        cursor: pointer;
+      }
+    }
+  }
+}
