+++ src/pages/jihwan/Main/CommentData.json
@@ -1,0 +1,14 @@
+[
+  {
+    "name": "사람1",
+    "comment": "comment1"
+  },
+  {
+    "name": "사람2",
+    "comment": "comment2"
+  },
+  {
+    "name": "사람3",
+    "comment": "comment3"
+  }
+]
