+++ src/pages/jihwan/Main/Main.js
@@ -1,31 +1,8 @@
-import { useState, React } from 'react';
+import React from 'react';
+import Comment from './Comment';
 import './Main.scss';
 
 const Main = () => {
-  const [comments, setComments] = useState([]);
-
-  const [commentMsg, setCommentMsg] = useState('');
-
-  const handleInputComment = e => {
-    setCommentMsg(e.target.value);
-    console.log(e.target.value);
-  };
-
-  const handleCommentSubmit = e => {
-    e.preventDefault();
-
-    if (commentMsg === '') return alert('댓글 내용을 입력해주세요');
-
-    setComments(prev => [...prev, commentMsg]);
-    return setCommentMsg('');
-  };
-
-  let activeSubmit = () => {
-    if (commentMsg === '') return true;
-  };
-
-  let registerBtnOn = activeSubmit() ? 'registerBtn' : 'registerBtnActive';
-
   return (
     <section className="main">
       <nav className="nav">
@@ -77,7 +54,11 @@ const Main = () => {
                 <span className="subTitle">weCode - 위코드</span>
               </div>
               <div>
-                <img className="dotImg" src="images/jihwan/dot.png" alt="dot" />
+                <img
+                  className="dotImg"
+                  src="/images/jihwan/dot.png"
+                  alt="dot"
+                />
               </div>
             </div>
             <div className="feedContainer">
@@ -86,71 +67,7 @@ const Main = () => {
                 src="images/jihwan/feedImg.png"
                 alt="feedImg"
               />
-              <div className="feedWrap">
-                <div className="feedSubImg">
-                  <img
-                    className="feedHeart"
-                    src="images/jihwan/heart.png"
-                    alt="feedHeart"
-                  />
-                  <img
-                    className="feedComment"
-                    src="images/jihwan/comment.png"
-                    alt="feedComment"
-                  />
-                  <img
-                    className="feedUpload"
-                    src="images/jihwan/upload.png"
-                    alt="feedUpload"
-                  />
-                  <img
-                    className="feedBookmark"
-                    src="images/jihwan/bookmark.png"
-                    alt="feedBookmark"
-                  />
-                </div>
-                <div className="contents">
-                  <img
-                    className="contentImg"
-                    src="images/jihwan/profile2.png"
-                    alt="contentImg"
-                  />
-                  <span className="spanBold">
-                    <span className="spBold1">ddd</span>님
-                    <span className="spBold2">외 11명</span>이 좋아합니다
-                  </span>
-                  <div className="contentSpanWrap">
-                    <ul>
-                      {comments.map(comment => {
-                        return (
-                          <li>
-                            {}
-                            {comment}
-                          </li>
-                        );
-                      })}
-                    </ul>
-                  </div>
-                  <form onSubmit={handleCommentSubmit}>
-                    <div className="comment">
-                      <input
-                        onChange={handleInputComment}
-                        value={commentMsg}
-                        className="commentInput"
-                        type="text"
-                        placeholder="댓글 달기..."
-                      />
-                      <button
-                        type="submit"
-                        className={registerBtnOn}
-                        onClick={handleCommentSubmit}
-                      >
-                        게시
-                      </button>
-                    </div>
-                  </form>
-                </div>
-              </div>
+              <Comment />
             </div>
           </div>
         </div>
